(*) python lick detection w/ adaptive threshold
1) python detects licks based on an adaptive threshold of the first
derivative. It looks at the 0.75 percentile of the non-zeros values and
scales accordingly. I want there to be a gui toggle to start and stop
this once a good value is locked in.

todo: I want the plots to be relative to actual time, but I am having
trouble indexing time on the fly.